[#182] feat: hub package 구조 domain별로 변경, collector module 부분의 consumerserviceImpl / consumerservice interface 구조로 변경
