+++ collector/src/main/java/com/wherecar/collector/gpslog/application/dto/GpsLogRequest.java
@@ -10,6 +10,7 @@ import java.util.List;
 @Builder
 @Getter
 @Setter
+@ToString
 @AllArgsConstructor
 @NoArgsConstructor
 public class GpsLogRequest {
@@ -30,17 +31,4 @@ public class GpsLogRequest {
     @JsonProperty("cList")
     private List<GpsLogInfo> cList = new ArrayList<>(); // 주기 정보 리스트
 
-    @Override
-    public String toString() {
-        return "GpsLogRequest{" +
-                "mdn='" + mdn + '\'' +
-                ", tid='" + tid + '\'' +
-                ", mid='" + mid + '\'' +
-                ", pv='" + pv + '\'' +
-                ", did='" + did + '\'' +
-                ", oTime='" + oTime + '\'' +
-                ", cCnt='" + cCnt + '\'' +
-                ", cList=" + cList +
-                '}';
-    }
 }
