--- hub/src/main/java/com/wherecar/hub/application/GpsLogHubService.java
@@ -1,8 +1,0 @@
-package com.wherecar.hub.application;
-
-import com.wherecar.hub.application.dto.GpsLogRequest;
-import com.wherecar.hub.application.dto.MessageResponse;
-
-public interface GpsLogHubService {
-    MessageResponse sendGpsLogMessage(GpsLogRequest gpsLogRequest);
-}
