+++ hub/src/main/java/com/wherecar/hub/carlog/application/dto/CarLogRequest.java
@@ -1,14 +1,12 @@
-package com.wherecar.hub.application.dto;
+package com.wherecar.hub.carlog.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 // 온 오프 로그 요청 폼
 @Getter
 @Builder
+@ToString
 @AllArgsConstructor
 @NoArgsConstructor
 public class CarLogRequest {
@@ -32,22 +30,4 @@ public class CarLogRequest {
     private String spd;     // 속도.           범위: 0 ~ 255(단위: km/h)
     private String sum;     // 누적 주행 거리.  범위: 0 ~ 9999999(단위: m)
 
-    @Override
-    public String toString() {
-        return "CarLogRequest{" +
-                "mdn='" + mdn + '\'' +
-                ", tid='" + tid + '\'' +
-                ", mid='" + mid + '\'' +
-                ", pv='" + pv + '\'' +
-                ", did='" + did + '\'' +
-                ", onTime='" + onTime + '\'' +
-                ", offTime='" + offTime + '\'' +
-                ", gcd='" + gcd + '\'' +
-                ", lat='" + lat + '\'' +
-                ", lon='" + lon + '\'' +
-                ", ang='" + ang + '\'' +
-                ", spd='" + spd + '\'' +
-                ", sum='" + sum + '\'' +
-                '}';
-    }
 }
