+++ hub/src/main/java/com/wherecar/hub/gpslog/application/GpsLogHubService.java
@@ -1,0 +1,8 @@
+package com.wherecar.hub.gpslog.application;
+
+import com.wherecar.hub.gpslog.application.dto.GpsLogRequest;
+import com.wherecar.hub.common.application.dto.MessageResponse;
+
+public interface GpsLogHubService {
+    MessageResponse sendGpsLogMessage(GpsLogRequest gpsLogRequest);
+}
